requirements.txt
# Direct dependencies of deb-pkg-tools.
cached-property >= 0.1.5
coloredlogs >= 0.4.6
executor >= 1.3
humanfriendly >= 2.1
python-debian >= 0.1.21-nmu2
six >= 1.10.0

# Undocumented transitive dependency of python-debian...
chardet
